Ticket page fixes and type Ticket
