+++ src/views/TicketPage.vue
@@ -2,86 +2,169 @@
   <ion-page>
     <ion-content>
       <div id="block">
-        <Tilt class="ticket" data-tilt data-tilt-full-page-listening>
-          <header class="ticket_wrapper">
-            <div class="ticket_header">
-
-              <div id="header_block">
-                <img src="../../public/assets/logos/logo-foresight-bk.png" alt="logo" height="100" paddingBottom="10" />
-                <!-- <qrcode-vue :value="qrcode"></qrcode-vue> -->
-                <img :src="qrResult" alt="QR Code" height="100" paddingBottom="10" />
-              </div>
-              <div id="header_block">
-                <h3>Voucher</h3> 🎟️#{{ (Math.random() * 1000000).toFixed(0) }}
-              </div>
-            </div>
-          </header>
-          <div class="ticket_divider">
-            <div class="ticket_notch"></div>
-            <div class="ticket_notch ticket_notch--right"></div>
-          </div>
-          <div class="ticket_body">
-            <section class="ticket_section">
-              <h3><strong>{{ ticket.firstName }} {{ ticket.lastName }}</strong></h3>
-              <p>{{ ticket.company }}</p>
-            </section>
-            <section class="ticket_section">
-              <h3><strong>Foresight 2022</strong></h3>
-              <p>5 Ottobre 2022 | ore 09:00</p>
-              <p>Auditorium Fondazione Cariplo</p>
-              <p>Largo Gustav Mahler, 20136 Milano MI</p>
-            </section>
-          </div>
-          <footer class="ticket_footer">
-            <img src="../../public/assets/logos/foresight-supporters.png" alt="believers" />
-          </footer>
-        </Tilt>
+        <SearchTicketVue v-if="!validEmail" class="cerca" />
+        <TicketTemplateVue :ticket="ticket" v-else />
+        <!-- <h4 id="believers" style="color: black" class="pad-20">Sound Design and live performance</h4> -->
+        <!-- <img src="../../public/assets/logos/orchestra.png" alt="believers" width="100" class="pad-20" /> -->
+        <h4 id="believers" style="color: black" class="pad-20">Our Believers</h4>
+        <img src="../../public/assets/logos/foresight-supporters1.png" alt="believers" width="700" class="pad-20" />
       </div>
-
     </ion-content>
   </ion-page>
-
 </template>
 
 <script lang="ts" setup>
-import {
-  IonContent,
-  IonPage,
-} from '@ionic/vue'
-import { ref } from 'vue'
+import { IonContent, IonPage } from '@ionic/vue'
+import { computed, reactive, ref, defineProps, onBeforeMount } from 'vue'
 import { MD5 } from 'crypto-js'
-import Tilt from 'vanilla-tilt-vue'
-import axios from 'axios';
+import axios from 'axios'
+import SearchTicketVue from './Ticket/SearchTicket.vue'
+import TicketTemplateVue from './Ticket/TicketTemplate.vue'
+import { Ticket } from '@/stores/guest/state'
+// import AddToCalendar from './components/AddToCalendar.vue'
+// import ManualAddGuest from './components/ManualAddGuest.vue'
 
-const ticket = ref({
-  id: MD5(window.location.href.split('/')[5] || '').toString(),
-  firstName: 'John',
-  lastName: 'Doe',
-  email: window.location.href.split('/')[5] || '',
-  company: 'Foresight',
+
+
+type Props = {
+  query?: string
+}
+
+const props = defineProps<Props>()
+
+
+const serverUrl = process.env.VUE_APP_SERVER_URL
+
+const params = computed(() => {
+  return {
+    email: props.query?.toLowerCase() || '',
+  }
 })
 
-axios.get(`/guests/${ticket.value.id}`).then((res) => {
-ticket.value.firstName = res.data.firstName
-ticket.value.lastName = res.data.lastName
-ticket.value.company = res.data.company
-}).catch((err) => {
-console.log(err)
+const validEmail = ref(false)
+
+
+const ticket = reactive<Ticket>({
+  id: MD5(params.value.email).toString(),
+  firstName: '',
+  lastName: '',
+  email: params.value.email,
+  company: '',
+  qrCode: '',
 })
 
-const qrResult = ref('')
-qrResult.value = `/guests/qr/${ticket.value.id}`
-const qrcode = ref('')
-qrcode.value = ticket.value.id
-console.log(qrcode.value)
+onBeforeMount(() => {
+  axios
+    .get(serverUrl + `api/guests/${params.value.email}/rsvp/`)
+    .then(response => {
+      ticket.firstName = response.data.firstName
+      ticket.lastName = response.data.lastName
+      ticket.company = response.data.companyName
+      ticket.id = response.data.emailHash
+      console.log('data:', response.data)
+      validEmail.value = true
+    })
+    .catch(error => {
+      console.log({ error })
+
+    })
+})
 
 </script>
 
 <style scoped>
+#container {
+  text-align: center;
+
+  position: absolute;
+  left: 0;
+  right: 0;
+  top: 50%;
+  transform: translateY(-50%);
+}
+
+#container strong {
+  font-size: 20px;
+  line-height: 26px;
+}
+
+#container p {
+  font-size: 16px;
+  line-height: 22px;
+
+  color: #8c8c8c;
+
+  margin: 0;
+}
+
+#container a {
+  text-decoration: none;
+}
+
+.input {
+  background: rgb(158, 200, 255, 0.3);
+  border-radius: 10px;
+  border: none;
+  color: black;
+  margin-bottom: 5px;
+  margin-top: 2px;
+}
+
+.login-logo {
+  padding: 20px 0;
+  min-height: 200px;
+  text-align: center;
+}
+
+.login-logo img {
+  max-width: 150px;
+}
+
+.cerca {
+  margin-top: 15px;
+  margin-bottom: 60px;
+}
+
+center {
+  background: white;
+  padding: 25px;
+  border-radius: 25px;
+  box-shadow: 0 10px 16px 0 rgb(0 0 0 / 20%), 0 6px 20px 0 rgb(0 0 0 / 19%) !important;
+}
+
+center strong {
+  font-size: 20px;
+  line-height: 26px;
+  color: black;
+}
+
 *,
 *::after {
   box-sizing: border-box;
   margin: 0;
+  /* text-shadow: 0px 0px 20px black; */
+  font-family: helvetica neue, helvetica, arial, verdana, sans-serif;
+}
+
+.title {
+  text-align: center;
+  font-size: 2rem;
+  margin: 1rem;
+  color: black;
+}
+
+.footer {
+  display: flex;
+  align-items: center;
+  margin: 1rem;
+  flex-direction: column;
+  align-content: stretch;
+  justify-content: space-between;
+}
+
+.btn {
+  background: #a23cfd;
+  --background: #a23cfd;
 }
 
 #block {
@@ -89,150 +172,359 @@ console.log(qrcode.value)
   align-items: center;
   justify-content: center;
   min-height: 100%;
-  color: #454f54;
-  background-color: #f4f5f6;
-  background-image: linear-gradient(to bottom left, #abb5ba, #d5dadd);
+  color: #999999;
+  background-color: whitesmoke;
+  flex-direction: column;
 }
 
-#header_block {
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 8px;
-  /* align-items: center;
-  flex-direction: column; */
+body {
+  background-color: #436ea5;
+}
 
+.pad-20 {
+  padding-top: 20px;
+}
+
+.mar-20 {
+  margin-top: 20px;
 }
 
 .ticket {
-  display: grid;
-  grid-template-rows: auto 1fr auto;
-  max-width: 24rem;
+  font-family: sans-serif;
+  background-repeat: no-repeat;
+  background-position: top;
+  background-size: 100%;
+  width: 700px;
+  height: 325px;
+  border-radius: 15px;
+  -webkit-filter: drop-shadow(1px 1px 3px rgba(0, 0, 0, 0.3));
+  filter: drop-shadow(1px 1px 3px rgba(0, 0, 0, 0.3));
+  display: inline-flex;
+  margin: 10% auto auto auto;
+  color: #fff;
+  border-radius: 20px;
+  margin-bottom: 25px;
 }
 
-.ticket_header,
-.ticket_body,
-.ticket_footer {
-  padding: 1.25rem;
-  background-color: white;
-  border: 1px solid #abb5ba;
-  box-shadow: 0 2px 4px rgba(41, 54, 61, 0.25);
+.details-block {
+  display: flex;
+  background-image: url('../../public/assets/media/bg-1.png');
+  border-radius: 20px;
+  flex-direction: column;
+  background-position: center;
+  background-size: cover;
+  background-repeat: repeat;
+  justify-content: space-between;
+  height: 100%;
+  width: 525px;
 }
 
-.ticket_header {
-  font-size: 1.5rem;
-  border-top: 0.25rem solid #4FF3FD;
-  border-bottom: none;
-  border-left: none;
-  border-right: none;
-  box-shadow: none;
-  animation: borderPulse 10s infinite;
-}
+.logos-block {
+  margin-top: 25px;
+  display: inline-flex;
 
-@keyframes borderPulse {
-  20% {
-    border-color: #bf59ee;
-  }
-
-  40% {
-    border-color: #4FF3FD;
-  }
-
-  60% {
-    border-color: #02D3DE;
-  }
-
-  80% {
-    border-color: #DBA3F5;
-  }
-
-  100% {
-    border-color: #4FF3FD;
-  }
-}
-
-.ticket_wrapper {
-  box-shadow: 0 2px 4px rgba(41, 54, 61, 0.25);
-  border-radius: 0.375em 0.375em 0 0;
-  overflow: hidden;
-}
-
-.ticket_divider {
-  position: relative;
-  height: 1rem;
-  background-color: white;
-  margin-left: 0.5rem;
-  margin-right: 0.5rem;
-}
-
-.ticket_divider::after {
-  content: "";
-  position: absolute;
-  height: 50%;
-  width: 100%;
-  top: 0;
-  border-bottom: 2px dashed #e9ebed;
-}
-
-.ticket_notch {
-  position: absolute;
-  left: -0.5rem;
-  width: 1rem;
-  height: 1rem;
-  overflow: hidden;
-}
-
-.ticket_notch::after {
-  content: "";
-  position: relative;
-  display: block;
-  width: 2rem;
-  height: 2rem;
-  right: 100%;
-  top: -50%;
-  border: 0.5rem solid white;
-  border-radius: 50%;
-  box-shadow: inset 0 2px 4px rgba(41, 54, 61, 0.25);
-}
-
-.ticket_notch--right {
-  left: auto;
-  right: -0.5rem;
-}
-
-.ticket_notch--right::after {
-  right: 0;
-}
-
-.ticket_body {
-  border-bottom: none;
-  border-top: none;
-}
-
-.ticket_body>*+* {
-  margin-top: 1.5rem;
-  padding-top: 1.5rem;
-  border-top: 1px solid #e9ebed;
-}
-
-.ticket_section>*+* {
-  margin-top: 0.25rem;
-}
-
-.ticket_section>h3 {
-  font-size: 1.125rem;
-  margin-bottom: 0.5rem;
-}
-
-.ticket_header,
-.ticket_footer {
-  font-weight: bold;
-  font-size: 1.25rem;
-  display: block;
   justify-content: space-between;
 }
 
-.ticket_footer {
-  border-top: 2px dashed #e9ebed;
-  border-radius: 0 0 0.325rem 0.325rem;
+.logo-img {
+  height: 100px;
+  margin-left: 35px;
+}
+
+.logo-lomb {
+  height: 35px;
+  margin-right: 20px;
+}
+
+.logo-lomb-mob {
+  display: none;
+}
+
+.guest {
+  font-size: 30px;
+  margin: 20px 25px 0 35px;
+  float: left;
+  font-weight: bold;
+}
+
+.name {
+  margin-top: 25px;
+  font-size: 30px;
+  text-transform: uppercase;
+}
+
+.company {
+  font-size: 20px;
+}
+
+.small {
+  font-weight: 200;
+}
+
+.location-block {
+  float: left;
+  margin: 25px 0 25px 35px;
+  font-size: 20px;
+  -webkit-text-stroke: 0.1px black;
+  font-weight: bold;
+  /* -webkit-filter: drop-shadow(1px 1px 3px rgba(0, 0, 0, 0.3));
+  filter: drop-shadow(1px 1px 3px rgba(0, 0, 0, 0.3)); */
+}
+
+.location-block::before {
+  background-size: 110px 110px;
+  width: 110px;
+  height: 110px;
+  content: '';
+  display: inline-block;
+  float: left;
+  position: absolute;
+  left: -160px;
+  bottom: 5px;
+  /* -webkit-filter: drop-shadow(1px 1px 3px rgba(0, 0, 0, 0.3));
+  filter: drop-shadow(1px 1px 3px rgba(0, 0, 0, 0.3)); */
+}
+
+.location {
+  font-size: 16px;
+  font-weight: 700;
+}
+
+.rip {
+  border-right: 8px dotted #436ea5;
+  height: 300px;
+  position: absolute;
+  top: 0;
+  left: 520px;
+}
+
+.qr-block {
+  width: 175px;
+  height: 100%;
+  background-color: #a23cfd;
+  border-radius: 20px;
+  color: #fff;
+  text-decoration: none;
+}
+
+.upper_block {
+  display: flex;
+  margin-top: 25px;
+  position: fixed;
+  justify-content: center;
+  align-items: center;
+  right: 25px;
+  flex-direction: column;
+  align-content: space-between;
+}
+
+.lower_block {
+  display: flex;
+  flex-wrap: nowrap;
+  flex-direction: column;
+  align-content: space-between;
+  justify-content: flex-end;
+  align-items: flex-end;
+  position: fixed;
+  bottom: 25px;
+  max-width: 175px;
+}
+
+.qr-img {
+  width: 125px;
+  margin-bottom: 15px;
+}
+
+.disclaimer {
+  font-size: 12px;
+  margin-left: 15px;
+  margin-right: 15px;
+  text-align: center;
+}
+
+h2#believers {
+  color: white;
+  /* padding-top: 20px; */
+  margin: 20px;
+  text-shadow: none;
+}
+
+/*  
+ Responsive design
+ */
+
+@media (max-width: 700px) {
+  .ticket {
+    font-family: sans-serif;
+    background-repeat: no-repeat;
+    background-position: top;
+    background-size: 100%;
+    width: 95%;
+    height: 680px;
+    border-radius: 15px;
+    filter: drop-shadow(1px 1px 3px rgba(0, 0, 0, 0.3));
+    display: flex;
+    margin: 10% auto auto auto;
+    color: #fff;
+    border-radius: 20px;
+    flex-direction: column;
+    flex-wrap: nowrap;
+    align-items: center;
+    margin-bottom: 25px;
+  }
+
+  .details-block {
+    display: flex;
+    background-image: url('../../public/assets/media/bg-1.png');
+    border-radius: 20px;
+    flex-direction: column;
+    background-position: center;
+    background-repeat: repeat;
+    background-size: cover;
+    justify-content: flex-start;
+    height: 75%;
+    width: 100%;
+    align-items: flex-start;
+  }
+
+  .logos-block {
+    margin-top: 25px;
+    display: inline-flex;
+
+    justify-content: space-between;
+  }
+
+  .logo-img {
+    height: 100px;
+    margin-left: 35px;
+  }
+
+  .logo-lomb {
+    height: 35px;
+    margin-right: 20px;
+    display: none;
+  }
+
+  .logo-lomb-mob {
+    display: block;
+    height: 35px;
+    margin-top: 60px;
+    margin-left: 35px;
+  }
+
+  .guest {
+    font-size: 25px;
+    margin: 20px 25px 0 35px;
+    float: left;
+    font-weight: bold;
+  }
+
+  .name {
+    margin-top: 25px;
+    font-size: 30px;
+    text-transform: uppercase;
+  }
+
+  .company {
+    font-size: 20px;
+  }
+
+  .small {
+    font-weight: 200;
+  }
+
+  .location-block {
+    float: left;
+    margin: 25px 35px 25px 35px;
+    font-size: 20px;
+    -webkit-text-stroke: 0.1px black;
+    font-weight: bold;
+    /* -webkit-filter: drop-shadow(1px 1px 3px rgba(0, 0, 0, 0.3));
+  filter: drop-shadow(1px 1px 3px rgba(0, 0, 0, 0.3)); */
+  }
+
+  .location-block::before {
+    background-size: 110px 110px;
+    width: 110px;
+    height: 110px;
+    content: '';
+    display: inline-block;
+    float: left;
+    position: absolute;
+    left: -160px;
+    bottom: 5px;
+    /* -webkit-filter: drop-shadow(1px 1px 3px rgba(0, 0, 0, 0.3));
+  filter: drop-shadow(1px 1px 3px rgba(0, 0, 0, 0.3)); */
+  }
+
+  .location {
+    font-size: 16px;
+    font-weight: 700;
+  }
+
+  .rip {
+    border-right: 8px dotted #436ea5;
+    height: 300px;
+    position: absolute;
+    top: 0;
+    left: 520px;
+  }
+
+  .qr-block {
+    width: 100%;
+    height: 25%;
+    background-color: #a23cfd;
+    border-radius: 20px;
+    color: #fff;
+    text-decoration: none;
+    display: flex;
+    flex-direction: row-reverse;
+    align-content: center;
+    justify-content: flex-start;
+    align-items: center;
+  }
+
+  .upper_block {
+    display: flex;
+    margin: 15px -15px 0px 15px;
+    width: 95%;
+    position: fixed;
+    justify-content: space-around;
+    align-items: flex-start;
+    flex-direction: row;
+    align-content: flex-start;
+    text-align: left;
+  }
+
+  .lower_block {
+    display: flex;
+    flex-wrap: nowrap;
+    flex-direction: column;
+    align-content: space-between;
+    justify-content: flex-end;
+    align-items: flex-end;
+    position: fixed;
+    bottom: 25px;
+    max-width: 175px;
+  }
+
+  .qr-img {
+    width: 125px;
+    margin-bottom: 15px;
+  }
+
+  .disclaimer {
+    font-size: 12px;
+    margin-left: 45px;
+    margin-right: 15px;
+    text-align: left;
+  }
+
+  h2#believers {
+    color: white;
+    /* padding-top: 20px; */
+    margin: 20px;
+    text-shadow: none;
+  }
 }
 </style>
